refactor modal into metadata
