+++ packages/apps-config/src/ui/index.ts
@@ -2,8 +2,16 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
+import { chainColors, nodeColors } from './general';
 import { identityNodes } from './identityIcons';
 
+function sanitize (value?: string): string {
+  return value?.toLowerCase().replace('-', ' ') || '';
+}
 export function getSystemIcon (systemName: string): 'beachball' | 'polkadot' | 'substrate' {
   return (identityNodes[systemName.toLowerCase().replace(/-/g, ' ')] || 'substrate') as 'substrate';
 }
+
+export const getSystemChainColor = (systemChain: string, systemName: string): string | undefined => {
+  return chainColors[sanitize(systemChain)] || nodeColors[sanitize(systemName)];
+};
