+++ packages/apps/public/locales/en/app-settings.json
@@ -16,7 +16,6 @@
   "default interface language": "default interface language",
   "Cancel": "Cancel",
   "Save & Reload": "Save & Reload",
-  "Export Network Specs": "Export Network Specs",
   "General": "General",
   "Developer": "Developer",
   "Select the remote endpoint, either from the dropdown on manual entered via the custom toggle": "Select the remote endpoint, either from the dropdown on manual entered via the custom toggle",
@@ -27,6 +26,7 @@
   "upgradable extensions": "upgradable extensions",
   "Update metadata": "Update metadata",
   "Extensions {{count}}": "Extensions {{count}}",
+  "Chain Specifications as a QR Code": "Chain Specifications as a QR Code",
   "Generate Random Color": "Generate Random Color",
   "Network Name": "Network Name",
   "Name of the network. It only for display purpose.": "Name of the network. It only for display purpose.",
