+++ packages/apps/public/locales/ja/app-settings.json
@@ -22,6 +22,9 @@
   "remote node/endpoint to connect to": "リモートノード・エンドポイント",
   "custom endpoint": "カスタムエンドポイント",
   "Default browser language (auto-detect)": "ブラウザー設定言語（自動反映）",
+  "Update metadata": "Update metadata",
+  "Extensions {{count}}": "Extensions {{count}}",
+  "Chain Specifications as a QR Code": "Chain Specifications as a QR Code",
   "Generate Random Color": "Generate Random Color",
   "Network Name": "Network Name",
   "Name of the network. It only for display purpose.": "Name of the network. It only for display purpose.",
