+++ packages/apps/public/locales/zh/app-settings.json
@@ -16,16 +16,17 @@
   "default interface language": "默认接口语言",
   "Cancel": "取消",
   "Save & Reload": "保存并重载",
-  "Export Network Specs": "导出网络配置",
   "General": "通用",
   "Developer": "开发者",
   "Select the remote endpoint, either from the dropdown on manual entered via the custom toggle": "选择远程终端，或者通过自定义切换输入的手动下拉菜单中的任意一个",
   "remote node/endpoint to connect to": "远程节点/终端连接到",
   "custom endpoint": "自定义终端",
   "Default browser language (auto-detect)": "默认浏览器语言 (自动检测)",
-  "No upgradable extensions found": "没有找到可以升级的浏览器插件",
+  "No upgradable extensions found": "没有需要升级的浏览器插件",
   "upgradable extensions": "升级浏览器插件",
-  "Update metadata": "升级metadata",
+  "Update metadata": "升级Metadata",
+  "Extensions {{count}}": "浏览器插件 {{count}}",
+  "Chain Specifications as a QR Code": "生成含网络参数的二维码",
   "Generate Random Color": "随机生成颜色",
   "Network Name": "网络名称",
   "Name of the network. It only for display purpose.": "网络名称，仅作为显示用",
