+++ packages/page-settings/src/Metadata/ChainColorIndicator.tsx
@@ -1,0 +1,23 @@
+// Copyright 2017-2020 @polkadot/app-settings authors & contributors
+// This software may be modified and distributed under the terms
+// of the Apache-2.0 license. See the LICENSE file for details.
+
+import React from 'react';
+import styled from 'styled-components';
+
+import { BareProps } from '@polkadot/react-components/types';
+
+interface Props extends BareProps {
+  color: string;
+}
+
+function ChainColorIndicator ({ color, className }: Props): React.ReactElement<Props> {
+  return <div color={color} className={className}/>;
+}
+
+export default React.memo(styled(ChainColorIndicator)`
+    background-color: ${(props: Props): string => props.color} !important;
+    width: 100px;
+    flex: 1;
+    border-radius: 4px;
+`);
