+++ packages/page-settings/src/index.tsx
@@ -4,7 +4,7 @@
 
 import { AppProps as Props } from '@polkadot/react-components/types';
 
-import React, { useMemo } from 'react';
+import React from 'react';
 import { Route, Switch } from 'react-router';
 import { HelpOverlay, Tabs } from '@polkadot/react-components';
 import uiSettings from '@polkadot/ui-settings';
@@ -12,7 +12,7 @@ import uiSettings from '@polkadot/ui-settings';
 import md from './md/basics.md';
 import { useTranslation } from './translate';
 import Developer from './Developer';
-import Extensions from './Extensions';
+import Metadata from './Metadata';
 import General from './General';
 import useCounter from './useCounter';
 
@@ -23,29 +23,22 @@ const hidden = uiSettings.uiMode === 'full'
   : ['developer'];
 
 function SettingsApp ({ basePath, onStatusChange }: Props): React.ReactElement<Props> {
-  const numExtensions = useCounter();
   const { t } = useTranslation();
-  const items = useMemo(() => [
+  const items = [
     {
       isRoot: true,
       name: 'general',
       text: t('General')
     },
     {
-      name: 'extensions',
-      text: t('Extensions {{count}}', {
-        replace: {
-          count: numExtensions
-            ? `(${numExtensions})`
-            : ''
-        }
-      })
+      name: 'metadata',
+      text: t('Metadata')
     },
     {
       name: 'developer',
       text: t('Developer')
     }
-  ], [numExtensions, t]);
+  ];
 
   return (
     <main className='settings--App'>
@@ -64,8 +57,8 @@ function SettingsApp ({ basePath, onStatusChange }: Props): React.ReactElement<P
             onStatusChange={onStatusChange}
           />
         </Route>
-        <Route path={`${basePath}/extensions`}>
-          <Extensions />
+        <Route path={`${basePath}/metadata`}>
+          <Metadata />
         </Route>
         <Route component={General} />
       </Switch>
